Updated to Nuget Mono.Cecil 0.10.4 & Nunit 3.12.0 // building against Unity 2019.1.10f1 (temp)
Updated to latest packages
Added a reference to Unity 2019.1.10f1 (hintpath refers to the windows install path with Hub)

Project compiles, but has 87 tests not passing due to, I suspect, either some of the NUnit code changes
or the way Cecil works now. That requires some debugging with IL viewing and such.